requirements.txt
dash==1.11.0
dash-bootstrap-components==0.9.2
dash-core-components==1.9.1
dash-html-components==1.0.3
dash-renderer==1.4.0
dash-table==4.6.2
gunicorn==20.0.4
numpy==1.18.1
pandas==1.0.3
pandas-datareader==0.8.1
plotly==4.6.0
requests==2.23.0
flask-caching-1.8.0
